Bind mount /etc inside the snapshot
If MicroOS is used, the /etc/crypttab version inside the snapshot /etc
can be a base that is actually modified via an overlay.  This is the
case when during the enrollment we add the use of TPM2.

When dracut is called via --sysroot, it will use the unmodified version
of /etc/crypttab, generating an initrd that will not contain the TPM2
libraries (/usr/lib64/cryptsetup/libcryptsetup-token-systemd-tpm2.so),
failling the open of the LUKS2 device.

This commit bind mount the external current /etc inside the snapshot,
making it visible for dracut to generate a current valid initrd.

Signed-off-by: Alberto Planas <aplanas@suse.com>