Call regenerate-initrd-posttrans
In MicroOS the initrd needs to be generated at the end of the
transaction.  This commit call the regenerate script from
suse-module-tools from the set_default_snapshot.

The renegerate script detect if the transaction is ongoing, and will
avoid the call to sdbootutil, but if the transaction is done (for
example, when called from the snapper hook) it will correctly generate a
new initrd.

The tukit plugin will migrate the /run/regenerate-initrd content (if
present) from inside the transaction to ouside, so the posttrans script
can decide over the initrd regeneration.

Signed-off-by: Alberto Planas <aplanas@suse.com>