Bind mount snapshot dir onto itself before chroot.
This allows dracut running inside chroot to properly detect the root
filesystem type and include the relevant dracut module.

If there is an x-initrd.mount option in the fstab entry for the rootfs
(seems to be the case for prebuilt images) this is not strictly needed,
but there isn't one if using .iso and the installer.