Measure GPT of the disk with ESP, not the disk with root
Firmware measures GPT of the disk where boot application was loaded from.
In practice it means ESP where we install our bootloader.

systemd-pcrlock defaults to the disk where root ("/") filesystem is
located. It can be anywhere and it can be on a virtual device, like
LUKS container or LVM, but it is not what firmware knows or measures.

Explicitly measure GPT with ESP to match the firmware behavior.

Fixes: #168