Determine if is transactional checking read-only-root-fs
To determine if we are in a transactional system we can check if fstab
is marking / as ro.  This works in MicroOS but not in Aeon.

The current check of non-writable /usr is also broken when the image is
building, as /usr can be writable for a period of time.

Checking if the read-only-root-fs via RPM mostly works, but has the risk
of interferring with the scriptlets.

A future configuration file, of a fix in the fstab marker (ro=vfs?) can
provide a more robust check.

Signed-off-by: Alberto Planas <aplanas@suse.com>