Fix boot entry detection with(out) boot counting
We will now check if an entry for the current snapshot already exists
(with out without boot counting) and determine based on the file hash
if we want to reuse it. This way, we do not reset boot counting if
the content didn't change, but also reset it if it changed.
We will, however, never remove the boot counter from existing entries,
even if disabled.